+++ Api/RA.Utilities.OpenApi/Models/OpenApiRequestExample.cs
@@ -1,0 +1,8 @@
+namespace RA.Utilities.OpenApi.Models;
+
+/// <summary>
+/// Represents an example for an OpenAPI request body.
+/// </summary>
+public class OpenApiRequestExample : OpenApiExampleModel
+{
+}